+++ src/index.ts
@@ -6,8 +6,10 @@ export { NearConnector } from "./NearConnector";
 
 export { nearActionsToConnectorActions } from "./actions";
 export type { ConnectorAction } from "./actions/types";
+export type { WalletPlugin } from "./types/plugin";
 
 export type {
+  FooterBranding,
   NearWalletBase,
   WalletManifest,
   EventNearWalletInjected,
@@ -16,5 +18,5 @@ export type {
   SignAndSendTransactionParams,
   SignAndSendTransactionsParams,
   SignDelegateActionsParams as SignDelegateActionParams,
-  SignDelegateActionResult
+  SignDelegateActionResult,
 } from "./types";