+++ src/popups/NearWalletsPopup.ts
@@ -1,6 +1,6 @@
 import { html } from "../helpers/html";
 import { parseUrl } from "../helpers/url";
-import { WalletManifest } from "../types";
+import { FooterBranding, WalletManifest } from "../types";
 import { Popup } from "./Popup";
 
 const debugManifest = {
@@ -18,7 +18,7 @@ const debugManifest = {
     signInWithoutAddKey: true,
     signAndSendTransaction: true,
     signAndSendTransactions: true,
-    signDelegateAction: true
+    signDelegateAction: true,
   },
   permissions: {
     storage: true,
@@ -30,6 +30,7 @@ export class NearWalletsPopup extends Popup<{ wallets: WalletManifest[]; showSet
   constructor(
     readonly delegate: {
       wallets: WalletManifest[];
+      footer: FooterBranding | null;
       onAddDebugManifest: (wallet: string) => Promise<WalletManifest>;
       onRemoveDebugManifest: (id: string) => Promise<void>;
       onSelect: (id: string) => void;
@@ -105,6 +106,18 @@ export class NearWalletsPopup extends Popup<{ wallets: WalletManifest[]; showSet
     `;
   }
 
+  get footer() {
+    if (!this.delegate.footer) return "";
+    const { icon, heading, link, linkText } = this.delegate.footer;
+    return html`
+      <div class="footer">
+        <img src="${icon}" alt="${heading}" />
+        <p>${heading}</p>
+        <a class="get-wallet-link" href="${link}" target="_blank">${linkText}</a>
+      </div>
+    `;
+  }
+
   get dom() {
     if (this.state.showSettings) {
       return html`
@@ -129,11 +142,7 @@ export class NearWalletsPopup extends Popup<{ wallets: WalletManifest[]; showSet
               <button class="add-debug-manifest-button">Add</button>
             </div>
 
-            <div class="footer">
-              <img src="https://tgapp.herewallet.app/images/hot/hot-icon.png" alt="HOT Connector" />
-              <p>HOT Connector</p>
-              <p class="get-wallet-link">Don't have a wallet?</p>
-            </div>
+            ${this.footer}
           </div>
         </div>
       `;
@@ -154,11 +163,7 @@ export class NearWalletsPopup extends Popup<{ wallets: WalletManifest[]; showSet
 
         <div class="modal-body">${this.state.wallets.map((wallet: WalletManifest) => this.walletDom(wallet))}</div>
 
-        <div class="footer">
-          <img src="https://tgapp.herewallet.app/images/hot/hot-icon.png" alt="HOT Connector" />
-          <p>HOT Connector</p>
-          <p class="get-wallet-link">Don't have a wallet?</p>
-        </div>
+        ${this.footer}
       </div>
     </div>`;
   }